# autoassist/apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoAssist+ API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local (Docker Compose)" },
    { "url": "http://localhost:3001", "description": "Local (node dev)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Estimate": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "itemsJson": { "type": "object", "additionalProperties": true },
          "laborJson": { "type": "object", "additionalProperties": true },
          "total": { "type": "number" },
          "currency": { "type": "string", "example": "UAH" },
          "approved": { "type": "boolean", "nullable": true },
          "validUntil": { "type": "string", "format": "date-time" }
        }
      },
      "CreateEstimateRequest": {
        "type": "object",
        "properties": {
          "orderId": { "type": "integer" },
          "providerId": { "type": "integer" },
          "lines": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "description": { "type": "string" },
                "amount": { "type": "number", "minimum": 0 }
              },
              "required": ["description","amount"]
            }
          },
          "total": { "type": "number", "minimum": 0 }
        },
        "required": ["orderId","providerId","lines","total"]
      },
      "UpdateEstimateRequest": {
        "type": "object",
        "properties": {
          "laborCost": { "type": "number" },
          "partsCost": { "type": "number" },
          "totalCost": { "type": "number" },
          "estimatedDays": { "type": "integer", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "breakdown": { "type": "object", "additionalProperties": true }
        }
      },
...
      "TowQuoteRequest": {
        "type": "object",
        "properties": {
          "orderId": { "type": "integer" },
          "pickup": {
            "type": "object",
            "properties": {
              "latitude": { "type": "number" },
              "longitude": { "type": "number" },
              "address": { "type": "string" }
            }
          },
          "destination": {
            "type": "object",
            "properties": {
              "latitude": { "type": "number" },
              "longitude": { "type": "number" },
              "address": { "type": "string" }
            }
          },
          "serviceLevel": { "type": "string", "enum": ["STANDARD","EXPRESS","PREMIUM"] }
        }
      },
...
      "TowAssignRequest": {
        "type": "object",
        "properties": {
          "towTruckId": { "type": "integer" },
          "driverId": { "type": "integer" },
          "serviceLevel": { "type": "string", "enum": ["STANDARD","EXPRESS","PREMIUM"] }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/estimates/{orderId}": {
      "get": {
        "summary": "Get estimate by orderId",
        "parameters": [{ "in": "path", "name": "orderId", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Estimate" } } } },
          "404": { "description": "Estimate not found" }
        }
      }
    },
    "/api/estimates": {
      "post": {
        "summary": "Create estimate",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateEstimateRequest" } } } },
        "responses": { "201": { "description": "Created" }, "409": { "description": "Already exists" } }
      }
    },
    "/api/estimates/{id}": {
      "put": {
        "summary": "Update estimate",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEstimateRequest" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete estimate",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "No Content" }, "404": { "description": "Not found" } }
      }
    },
    "/api/estimates/{id}/approve": {
      "post": {
        "summary": "Approve estimate",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
    "/api/estimates/{id}/reject": {
      "post": {
        "summary": "Reject estimate",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "reason": { "type": "string" } } } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
...
    "/api/tow/quote": {
      "post": {
        "summary": "Create/Get tow quote",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TowQuoteRequest" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/tow/{orderId}/assign": {
      "post": {
        "summary": "Assign tow truck to order",
        "parameters": [{ "in": "path", "name": "orderId", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TowAssignRequest" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Tow request not found" } }
      }
    },
...
    "/api/notifications/test": {
      "post": {
        "summary": "Send test notification (admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "userId": { "type": "integer" }, "type": { "type": "string" }, "title": { "type": "string" }, "message": { "type": "string" }, "priority": { "type": "string" }, "channels": { "type": "array", "items": { "type": "string" } } } }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" } }
      }
    },
...
    "/api/orders/{id}/complete": {
      "post": {
        "summary": "Complete order and record proof-hash",
        "parameters": [ { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "photos": { "type": "array", "items": { "type": "integer" } },
                  "coords": { "type": "object", "properties": { "lat": { "type": "number" }, "lng": { "type": "number" } } },
                  "completedAt": { "type": "string", "format": "date-time" },
                  "notes": { "type": "string" }
                },
                "required": ["photos"]
              }
              ,
              "examples": {
                "done": {
                  "summary": "Completion with geo and photos",
                  "value": {
                    "photos": [654, 655],
                    "coords": { "lat": 50.4501, "lng": 30.5234 },
                    "completedAt": "2025-10-22T12:30:00Z",
                    "notes": "Client confirmed pick-up"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order completed",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "order": { "type": "object", "properties": { "id": { "type": "integer" }, "status": { "type": "string" } } }, "proofHash": { "type": "string" } }, "required": ["order","proofHash"] },
                "examples": { "ok": { "value": { "order": { "id": 142, "status": "CLOSED" }, "proofHash": "sha256:ef2a7f0a9d4a..." } } }
              }
            }
          },
          "400": { "description": "Invalid id" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/orders/{id}/proof": {
      "get": {
        "summary": "Get completion proof for order",
        "parameters": [ { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": {
            "description": "Proof details",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "orderId": { "type": "integer" }, "proofHash": { "type": "string" }, "evidence": { "type": "object" }, "createdAt": { "type": "string", "format": "date-time" } }, "required": ["orderId","proofHash","createdAt"] } } }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
  },
...
  }
}